Refactored Layout of Video Page and Intro
