--- src/Layouts/Main/TechInDailyLife/TechInDailyLifeintro/index.module.css
@@ -1,8 +1,0 @@
-.homeContainer {
-	margin-top: 2rem;
-}
-
-.titleBox {
-	margin: auto;
-	width: 100%;
-}
