+++ src/Layouts/Main/TechVideos/index.jsx
@@ -1,0 +1,89 @@
+import React, { useEffect, useState } from 'react';
+import { Box } from '@mui/material';
+import { Breadcrumb } from '../../../components/Video/Breadcrumb';
+import FilterPanel from '../../../components/FilterPanel';
+import Searchbar from '../../../components/Video/Searchbar';
+import SubtopicSelection from '../../../components/Video/SubtopicSelection';
+import YouTubeVideoSection from '../../../components/Video/YouTubeVideoSection';
+import { collection, onSnapshot } from 'firebase/firestore';
+import { db } from '../../../firebase/firebase';
+import Intro from '../../../components/Video/Intro';
+
+function VideoLayout({ subtoptics, pageValue, introText }) {
+
+	useEffect(() => {
+		console.log('useEffect 1');
+		const unsubscribe = onSnapshot(docRef, (querySnapshot) => {
+			const videos = querySnapshot.docs.map((doc) => doc.data());
+			setDatafromFirebase(Object.values(videos));
+		});
+
+		return unsubscribe;
+		// eslint-disable-next-line
+	}, []);
+
+	const dataFromIntro = (osvalue) => {
+		setosValue(osvalue);
+	};
+
+	const [osvalue, setosValue] = useState([]);
+	const [dataFromFirebase, setDatafromFirebase] = useState([]);
+	const docRef = collection(db, 'youtube-videos');
+
+	// video search constants
+	const [subtopicValue, setsubtopicValue] = useState([]);
+	const [tags, tagsFromSearchBar] = useState([]);
+
+	const dataFromSubtopicSelector = (subtopicValue) => {
+		setsubtopicValue(subtopicValue);
+	};
+
+	const handleResetSubtopic = () => {
+		setsubtopicValue([]);
+	};
+
+	return (
+		<div>
+			<FilterPanel
+				filterGroups={[
+					{
+						subheading: 'Device Type',
+						filters: ['Mobile - iOS', 'Mobile - Android', 'Desktop - Windows', 'Desktop - Mac', 'Desktop - Linux'],
+					},
+					{
+						subheading: 'Content Type',
+						filters: ['Daily Life', 'Finance', 'Safety Privacy'],
+					},
+				]}
+			/>
+
+			<div className="md:pl-80">
+				<Intro
+					dataFromIntro={dataFromIntro}
+					dataFromFirebase={dataFromFirebase}
+					pageValue={pageValue}
+					introText={introText}
+				/>
+
+				<Box style={{ margin: 'auto', width: '70%', paddingBottom: '2rem' }}>
+					<Searchbar tagsFromSearchBar={tagsFromSearchBar} tags={tags} />
+					<Breadcrumb subtopicValue={subtopicValue} handleResetSubtopic={handleResetSubtopic} subtopics={subtoptics} />
+				</Box>
+
+				{subtoptics.length == 0 || subtopicValue.length > 0 || tags.length > 0 ? (
+					<YouTubeVideoSection
+						osvalue={osvalue}
+						subtopicValue={subtopicValue}
+						tags={tags}
+					/>
+				) : (
+					<SubtopicSelection
+						dataFromSubtopicSelector={dataFromSubtopicSelector}
+						subtopics={subtoptics} />
+				)}
+			</div>
+		</div>
+	);
+}
+
+export default VideoLayout;
